--- front/src/components/MainAboutUs.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const MainAboutUs = () => {
-  return (
-    <div>MainAboutUs</div>
-  )
-}
-
-export default MainAboutUs
