+++ front/src/pages/AboutUs.jsx
@@ -1,15 +1,8 @@
 import React from 'react'
-import Header from '../components/Header'
-import MainAboutUs from '../components/MainAboutUs'
-import Footer from '../components/Footer'
 
 const AboutUs = () => {
   return (
-    <div>
-      
-      <MainAboutUs/>
-      
-    </div>
+    <div>MainAboutUs</div>
   )
 }
